Factor out `AnnotationSyncService`
Factor the `SearchIndexService.sync()` method and its helpers out into a
separate `AnnotationSyncService`.

The concept of syncing annotations from Postgres to Elasticsearch is
distinct and complex enough to deserve its own separate service. This'll
make future PRs that aim to refactor and extend the `sync()` method
easier.
